requirements.txt
discord.py
python-dotenv
aiohttp
pytz
dateparser
beautifulsoup4
python-dateutil
flask
feedparser